Add target ratio categories.
This will allow you to easily switch ETFs if a cheaper one comes out.
All positions under a category will be combined and the first
in the list will be bought and the last will be sold. (When selling is added)

This changes the target_json format.

Added Canadian Couch Potato and Canadian portfolio Manager sample portfolios.